small changes to remove warnings on newer versions of GCC, and minor updates to windows build config file

+++ src/mrview/sidebar/tractography/track_list.cpp
@@ -398,7 +398,11 @@ namespace MR {
         Gtk::ColorSelectionDialog dialog ("Choose colour for ROI");
         if (dialog.run() == Gtk::RESPONSE_OK) {
           Gdk::Color colour (dialog.get_colorsel()->get_current_color());
-          GLubyte C[] = { colour.get_red() >> 8, colour.get_green() >> 8, colour.get_blue() >> 8 };
+          GLubyte C[] = { 
+            GLubyte(colour.get_red() >> 8),
+            GLubyte(colour.get_green() >> 8), 
+            GLubyte(colour.get_blue() >> 8)
+          };
           std::list<Gtk::TreeModel::Path> paths = get_selection()->get_selected_rows();
           for (std::list<Gtk::TreeModel::Path>::iterator i = paths.begin(); i != paths.end(); ++i) {
             Gtk::TreeModel::iterator iter = model->get_iter (*i);
